# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Top deal shop dự án tốt nghiệp.",
        "description": "https://apitopdeal.shop/",
        "contact": {
            "name": "Tên người liên hệ",
            "email": "email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/categories": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Thêm danh mục mới",
                "description": "Thêm danh mục vào hệ thống.",
                "operationId": "d2b35462e503ce58ae15900db673ef51",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Tên danh mục",
                                        "type": "string",
                                        "example": "Danh mục 1"
                                    },
                                    "image": {
                                        "description": "Hình ảnh danh mục",
                                        "type": "string",
                                        "example": "image.png"
                                    },
                                    "parent_id": {
                                        "description": "ID danh mục cha",
                                        "type": "bigint",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thêm danh mục thành công."
                                        },
                                        "data": {
                                            "properties": {
                                                "category": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "bigint",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Danh mục 1"
                                                        },
                                                        "image_url": {
                                                            "type": "string",
                                                            "example": "https://apitopdeal.shop/storage/categories/image.png"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Thêm danh mục thất bại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thêm danh mục thất bại"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error details..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "api/admin/categories/{id}": {
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Xóa danh mục",
                "operationId": "8ef4fc1e69fa1a5c85d87ec3f045cad4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục cần xóa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa danh mục thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Xóa danh mục thành công."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi khi xóa danh mục",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Xóa danh mục không thành công."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Lý do lỗi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Nhập token JWT trong định dạng 'Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}